Cases/text-classification/imdb-movies-genre/requirements.txt
nltk
contractions
numpy
pandas
pyarrow
scikit-learn
matplotlib
seaborn
pydantic
pyyaml
